splitting up validate module + adding cli command
